+++ Products/GenericSetup/tool.py
@@ -58,6 +58,13 @@ from Products.GenericSetup.utils import _wwwdir
 IMPORT_STEPS_XML = 'import_steps.xml'
 EXPORT_STEPS_XML = 'export_steps.xml'
 TOOLSET_XML = 'toolset.xml'
+# What do we do with already applied dependency profiles?
+# The behavior used to be: reapply.  Now there are more options.
+DEPENDENCY_STRATEGY_UPGRADE = 'upgrade'
+DEPENDENCY_STRATEGY_REAPPLY = 'reapply'
+DEPENDENCY_STRATEGY_NEW = 'new'
+DEPENDENCY_STRATEGY_IGNORE = 'ignore'
+DEFAULT_DEPENDENCY_STRATEGY = DEPENDENCY_STRATEGY_UPGRADE
 
 generic_logger = logging.getLogger(__name__)
 
@@ -352,7 +359,8 @@ class SetupTool(Folder):
                                      purge_old=None,
                                      ignore_dependencies=False,
                                      archive=None,
-                                     blacklisted_steps=None):
+                                     blacklisted_steps=None,
+                                     dependency_strategy=None):
         """ See ISetupTool.
         """
         __traceback_info__ = profile_id
@@ -362,7 +370,8 @@ class SetupTool(Folder):
                             profile_id=profile_id,
                             archive=archive,
                             ignore_dependencies=ignore_dependencies,
-                            blacklisted_steps=blacklisted_steps)
+                            blacklisted_steps=blacklisted_steps,
+                            dependency_strategy=dependency_strategy)
         if profile_id is None:
             prefix = 'import-all-from-tar'
         else:
@@ -536,12 +545,14 @@ class SetupTool(Folder):
                                        messages=messages)
 
     security.declareProtected(ManagePortal, 'manage_importAllSteps')
-    def manage_importAllSteps(self, context_id=None):
+    def manage_importAllSteps(self, context_id=None, dependency_strategy=None):
         """ Import all steps.
         """
         if context_id is None:
             context_id = self.getBaselineContextID()
-        result = self.runAllImportStepsFromProfile(context_id, purge_old=None)
+        result = self.runAllImportStepsFromProfile(
+            context_id, purge_old=None,
+            dependency_strategy=dependency_strategy)
 
         steps_run = 'Steps run: %s' % ', '.join(result['steps'])
 
@@ -803,6 +814,9 @@ class SetupTool(Folder):
     def getLastVersionForProfile(self, profile_id):
         """Return the last upgraded version for the specified profile.
         """
+        prefix = 'profile-'
+        if profile_id.startswith(prefix):
+            profile_id = profile_id[len(prefix):]
         version = self._profile_upgrade_versions.get(profile_id, 'unknown')
         return version
 
@@ -810,6 +824,9 @@ class SetupTool(Folder):
     def setLastVersionForProfile(self, profile_id, version):
         """Set the last upgraded version for the specified profile.
         """
+        prefix = 'profile-'
+        if profile_id.startswith(prefix):
+            profile_id = profile_id[len(prefix):]
         if isinstance(version, basestring):
             version = tuple(version.split('.'))
         prof_versions = self._profile_upgrade_versions.copy()
@@ -835,10 +852,6 @@ class SetupTool(Folder):
 
     security.declareProtected(ManagePortal, "getProfileInfo")
     def getProfileInfo(self, profile_id):
-        if profile_id.startswith("profile-"):
-            profile_id = profile_id[len('profile-'):]
-        elif profile_id.startswith("snapshot-"):
-            profile_id = profile_id[len('snapshot-'):]
         return _profile_registry.getProfileInfo(profile_id)
 
     security.declareProtected(ManagePortal, 'getDependenciesForProfile')
@@ -990,30 +1003,35 @@ class SetupTool(Folder):
     security.declarePrivate('_getImportContext')
     def _getImportContext(self, context_id, should_purge=None, archive=None):
         """ Crack ID and generate appropriate import context.
+
+        Note: it seems context_id (profile id) and archive (tarball)
+        are mutually exclusive.  Exactly one of the two should be
+        None.  There seems to be no use case for a different
+        combination.
         """
         encoding = self.getEncoding()
 
         if context_id is not None:
-            if context_id.startswith('profile-'):
-                context_id = context_id[len('profile-'):]
-                info = _profile_registry.getProfileInfo(context_id)
-
-                if info.get('product'):
-                    path = os.path.join(_getProductPath(info['product']),
-                                        info['path'])
-                else:
-                    path = info['path']
-                if should_purge is None:
-                    should_purge = (info.get('type') != EXTENSION)
-                return DirectoryImportContext(self, path, should_purge,
-                                              encoding)
-
-            elif context_id.startswith('snapshot-'):
-                context_id = context_id[len('snapshot-'):]
+            prefix = 'snapshot-'
+            if context_id.startswith(prefix):
+                context_id = context_id[len(prefix):]
                 if should_purge is None:
                     should_purge = True
                 return SnapshotImportContext(self, context_id, should_purge,
                                              encoding)
+            prefix = 'profile-'
+            if context_id.startswith(prefix):
+                context_id = context_id[len(prefix):]
+            info = _profile_registry.getProfileInfo(context_id)
+            if info.get('product'):
+                path = os.path.join(_getProductPath(info['product']),
+                                    info['path'])
+            else:
+                path = info['path']
+            if should_purge is None:
+                should_purge = (info.get('type') != EXTENSION)
+            return DirectoryImportContext(self, path, should_purge,
+                                          encoding)
 
         if archive is not None:
             return TarballImportContext(tool=self,
@@ -1146,8 +1164,46 @@ class SetupTool(Folder):
                                    profile_id=None,
                                    archive=None,
                                    ignore_dependencies=False,
-                                   seen=None,
-                                   blacklisted_steps=None):
+                                   blacklisted_steps=None,
+                                   dependency_strategy=None):
+
+        # 1. Determine upgrade strategy.
+        #    What do we do with already applied dependency profiles?
+
+        # There are two ways to say you want to ignore all
+        # dependencies.  If one is enabled, we enable the other too.
+        if dependency_strategy == DEPENDENCY_STRATEGY_IGNORE:
+            ignore_dependencies = True
+        elif ignore_dependencies:
+            dependency_strategy = DEPENDENCY_STRATEGY_IGNORE
+        # Turn None into the default:
+        if dependency_strategy is None:
+            dependency_strategy = DEFAULT_DEPENDENCY_STRATEGY
+        # Determine the settings based on the strategy.
+        if dependency_strategy == DEPENDENCY_STRATEGY_UPGRADE:
+            apply_new_profiles = True
+            reapply_old_profiles = False
+            upgrade_old_profiles = True
+        elif dependency_strategy == DEPENDENCY_STRATEGY_REAPPLY:
+            apply_new_profiles = True
+            reapply_old_profiles = True
+            upgrade_old_profiles = False
+        elif dependency_strategy == DEPENDENCY_STRATEGY_NEW:
+            apply_new_profiles = True
+            reapply_old_profiles = False
+            upgrade_old_profiles = False
+        elif dependency_strategy == DEPENDENCY_STRATEGY_IGNORE:
+            apply_new_profiles = False
+            reapply_old_profiles = False
+            upgrade_old_profiles = False
+        else:
+            raise ValueError('Unknown dependency_strategy %r.' %
+                             dependency_strategy)
+        generic_logger.info(
+            'Importing profile %s with dependency strategy %s.',
+            profile_id, dependency_strategy)
+
+        # 2. Gather a list of profiles to handle.
 
         if profile_id is not None and not ignore_dependencies:
             try:
@@ -1157,24 +1213,52 @@ class SetupTool(Folder):
                 logger.error('Unknown step in dependency chain: %s' % str(e))
                 raise
         else:
+            # Two possibilities:
+            # - We ignore dependencies, so we have a single profile id.
+            # - Profile id is None and we import a tarball (in the archive).
             chain = [profile_id]
-            if seen is None:
-                seen = set()
-            seen.add(profile_id)
+
+        # 3. For each profile, depending on the keyword arguments, either:
+        # a. do nothing or
+        # b. apply its upgrade steps or
+        # c. apply its full profile.
 
         results = []
-
         detect_steps = steps is None
 
-        for profile_id in chain:
+        # The chain is: first all dependency profiles ( recursively if
+        # applicable), and as last one the main profile for which we
+        # got passed the profile_id.
+        last_num = len(chain)
+        for num, profile_id in enumerate(chain, 1):
+            if num == last_num:
+                generic_logger.info('Applying main profile %s', profile_id)
+            else:
+                # This is a dependency profile.  This means we are not
+                # completely ignoring them, otherwise it would not
+                # have ended up in the list.  Check if it was already
+                # applied.
+                if self.getLastVersionForProfile(profile_id) == 'unknown':
+                    # This is a new profile.
+                    if not apply_new_profiles:
+                        continue
+                    generic_logger.info('Applying profile %s', profile_id)
+                else:
+                    # Profile was already applied.
+                    if upgrade_old_profiles:
+                        self.upgradeProfile(profile_id)
+                        continue
+                    if not reapply_old_profiles:
+                        continue
+                    generic_logger.info('Reapplying profile %s', profile_id)
+            # The next lines are done at least for the main profile.
+            # Possibly also for dependency profiles, depending on the
+            # condition above.  It applies the profile.
             context = self._getImportContext(profile_id, purge_old, archive)
             self.applyContext(context)
-
             if detect_steps:
                 steps = self.getSortedImportSteps()
-
             messages = {}
-
             event.notify(
                 BeforeProfileImportEvent(self, profile_id, steps, True))
             for step in steps:
@@ -1189,8 +1273,11 @@ class SetupTool(Folder):
                 context.clearNotes()
 
             event.notify(ProfileImportedEvent(self, profile_id, steps, True))
-
+            messages[profile_id] = (
+                'Imported with dependency strategy %s.' % dependency_strategy)
             results.append({'steps': steps, 'messages': messages})
+
+        # 4. Gather data for reporting back.
 
         data = {'steps': [], 'messages': {}}
         for result in results: